requirements.txt
Django==1.6
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
facebook-sdk==0.4.0
pyquery==1.2.1
requests==1.2.3
icalendar==3.6.1